backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.101.1
uvicorn[standard]==0.22.0
pydantic==2.5.1

# Database dependencies
SQLAlchemy==1.4.49
asyncpg==0.27.0
alembic==1.11.1
psycopg2-binary==2.9.7

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email validation
email-validator==2.0.0

# Additional utilities
python-decouple==3.8
Pillow==10.0.1
python-dateutil==2.8.2

# Development dependencies (optional)
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
